refactor : separate directory by function
