+++ src/main/java/org/example/board_project/exception/ErrorCode.java
@@ -14,7 +14,11 @@ public enum ErrorCode {
     FAIL_TO_UPLOAD_BOARD(HttpStatus.SERVICE_UNAVAILABLE, "FAILED UPLOAD BOARD"),
     FAIL_TO_UPDATE_BOARD(HttpStatus.SERVICE_UNAVAILABLE, "FAILED TO UPDATE BOARD"),
     FAIL_TO_DELETE_BOARD(HttpStatus.SERVICE_UNAVAILABLE, "FAILED DELETE BOARD"),
-    DTO_OBJECT_IS_EMPTY(HttpStatus.BAD_REQUEST, "DTO OBJECT IS EMPTY");
+    DTO_OBJECT_IS_EMPTY(HttpStatus.BAD_REQUEST, "DTO OBJECT IS EMPTY"),
+
+
+    NONEXISTENT_CATEGORY(HttpStatus.NOT_FOUND, "NONEXISTENT_CATEGORY");
+
 
     private final HttpStatus httpStatus;
     private final String message;
