+++ src/main/java/org/example/board_project/mapper/BoardMapper.java
@@ -3,7 +3,7 @@ package org.example.board_project.mapper;
 import org.apache.ibatis.annotations.Mapper;
 import org.example.board_project.model.dto.requestDTO.board.BoardListRequestDTO;
 import org.example.board_project.model.dto.requestDTO.board.WriteBoardRequestDTO;
-import org.example.board_project.model.entity.Board;
+import org.example.board_project.model.Board;
 
 import java.util.List;
