+++ src/test/java/org/example/board_project/boardTest/BoardServiceTest.java
@@ -4,8 +4,8 @@ package org.example.board_project.boardTest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.example.board_project.mapper.BoardMapper;
 import org.example.board_project.model.dto.responseDTO.board.BoardResponseDTO;
-import org.example.board_project.model.entity.Board;
-import org.example.board_project.service.BoardService;
+import org.example.board_project.model.Board;
+import org.example.board_project.service.board.BoardService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
